# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HVAC Training App API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Progress {
  id: number;
  moduleId: string;
  moduleName: string;
  completed: boolean;
  completedAt?: string | null;
  score?: number | null;
}

export interface SaveProgressInput {
  moduleId: string;
  moduleName: string;
  completed: boolean;
  score?: number | null;
}

export interface QuizResult {
  id: number;
  quizId: string;
  quizName: string;
  score: number;
  totalQuestions: number;
  completedAt: string;
}

export interface QuizResultInput {
  quizId: string;
  quizName: string;
  score: number;
  totalQuestions: number;
}