# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: HVAC Training App API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /quiz/results:
    post:
      operationId: submitQuizResult
      tags: [quiz]
      summary: Submit a quiz result
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizResultInput"
      responses:
        "200":
          description: Saved quiz result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizResult"
    get:
      operationId: getQuizResults
      tags: [quiz]
      summary: Get all quiz results
      responses:
        "200":
          description: List of quiz results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QuizResult"
components:
  schemas:
...
    QuizResult:
      type: object
      properties:
        id:
          type: integer
        quizId:
          type: string
        quizName:
          type: string
        score:
          type: integer
        totalQuestions:
          type: integer
        completedAt:
          type: string
          format: date-time
      required:
        - id
        - quizId
        - quizName
        - score
        - totalQuestions
        - completedAt
    QuizResultInput:
      type: object
      properties:
        quizId:
          type: string
        quizName:
          type: string
        score:
          type: integer
        totalQuestions:
          type: integer
      required:
        - quizId
        - quizName
        - score
        - totalQuestions