# lib/api-zod/src/generated/types/quizResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HVAC Training App API
 * OpenAPI spec version: 0.1.0
 */

export interface QuizResult {
  id: number;
  quizId: string;
  quizName: string;
  score: number;
  totalQuestions: number;
  completedAt: Date;
}